Reorganize editor-core package
