+++ packages/editor-core/src/char/__tests__/util.test.ts
