+++ packages/editor-core/src/char/types.ts
