+++ packages/editor-core/src/char/util.ts
