+++ packages/editor-core/src/printer/printer.ts
@@ -3,8 +3,8 @@
  */
 import * as Semantic from "@math-blocks/semantic";
 
-import * as types from "../ast/types";
-import * as builders from "../ast/builders";
+import * as types from "../char/types";
+import * as builders from "../char/builders";
 
 // TODO: when parsing editor nodes provide some way to link to the IDs of
 // the original nodes, even if they don't appear in the semantic tree as
