+++ packages/editor-core/src/reducer/__tests__/move-cursor.test.ts
@@ -1,5 +1,5 @@
-import * as types from "../../ast/types";
-import * as builders from "../../ast/builders";
+import * as types from "../../char/types";
+import * as builders from "../../char/builders";
 
 import {moveLeft} from "../move-left";
 import {moveRight} from "../move-right";
