+++ packages/editor-core/src/reducer/cancel.ts
@@ -1,4 +1,4 @@
-import * as transforms from "../ast/transforms";
+import * as transforms from "../char/transforms";
 import {selectionZipperFromZippers} from "./convert";
 
 import type {State} from "./types";
