+++ packages/editor-core/src/reducer/matrix.ts
@@ -1,7 +1,7 @@
 import type {Mutable} from "utility-types";
 
-import * as builders from "../ast/builders";
-import * as types from "../ast/types";
+import * as builders from "../char/builders";
+import * as types from "../char/types";
 
 import type {State, Zipper, Breadcrumb, ZTable, ZRow} from "./types";
