+++ packages/editor-core/src/reducer/parens.ts
@@ -1,5 +1,5 @@
-import * as builders from "../ast/builders";
-import * as types from "../ast/types";
+import * as builders from "../char/builders";
+import * as types from "../char/types";
 
 import {moveRight} from "./move-right";
