+++ packages/editor-core/src/reducer/test-util.ts
@@ -1,7 +1,7 @@
 import * as Semantic from "@math-blocks/semantic";
 
-import * as types from "../ast/types";
-import * as builders from "../ast/builders";
+import * as types from "../char/types";
+import * as builders from "../char/builders";
 
 import type {ZRow, Zipper, State} from "./types";
