+++ packages/editor-core/src/reducer/vertical-work/backspace.ts
@@ -1,6 +1,6 @@
-import * as types from "../../ast/types";
-import * as builders from "../../ast/builders";
-import {isAtom} from "../../ast/util";
+import * as types from "../../char/types";
+import * as builders from "../../char/builders";
+import {isAtom} from "../../char/util";
 
 import * as util from "../util";
 import {cursorLeft} from "../move-left";
