+++ packages/editor-core/src/reducer/vertical-work/util.ts
@@ -1,7 +1,7 @@
 import {getId} from "@math-blocks/core";
-import * as types from "../../ast/types";
-import * as builders from "../../ast/builders";
-import * as util from "../../ast/util";
+import * as types from "../../char/types";
+import * as builders from "../../char/builders";
+import * as util from "../../char/util";
 import type {ZTable, Zipper, Focus} from "../types";
 import {zrowToRow, zrow} from "../util";
 import type {Column, VerticalWork} from "./types";
