+++ packages/editor-core/src/token/serializer.ts
@@ -1,76 +1,6 @@
 import {UnreachableCaseError} from "@math-blocks/core";
 
-import {locFromRange} from "./util";
-import {
-    Token,
-    TokenNode,
-    TokenRow,
-    TokenSubSup,
-    TokenFrac,
-    TokenRoot,
-    TokenAtom,
-    SourceLocation,
-} from "./types";
-
-export function row(children: readonly TokenNode[]): TokenRow {
-    // What should the location be for an empty row?
-    const loc =
-        children.length > 0
-            ? locFromRange(children[0].loc, children[children.length - 1].loc)
-            : undefined;
-
-    return {
-        type: "row",
-        children,
-        // TODO: fix the path
-        loc: loc || {path: [], start: -1, end: -1},
-    };
-}
-
-export function subsup(
-    sub: readonly TokenNode[] | void,
-    sup: readonly TokenNode[] | void,
-    loc: SourceLocation,
-): TokenSubSup {
-    return {
-        type: "subsup",
-        children: [sub ? row(sub) : null, sup ? row(sup) : null],
-        loc,
-    };
-}
-
-export function frac(
-    numerator: readonly TokenNode[],
-    denominator: readonly TokenNode[],
-    loc: SourceLocation,
-): TokenFrac {
-    return {
-        type: "frac",
-        children: [row(numerator), row(denominator)],
-        loc,
-    };
-}
-
-// It would be nice if we could provide defaults to parameterized functions
-// We'd need type-classes for that but thye don't exist in JavaScript.
-export function root(
-    radicand: readonly TokenNode[],
-    index: readonly TokenNode[] | null,
-    loc: SourceLocation,
-): TokenRoot {
-    return {
-        type: "root",
-        children: [index ? row(index) : null, row(radicand)],
-        loc,
-    };
-}
-
-export function atom(token: Token, loc: SourceLocation): TokenAtom {
-    return {
-        ...token,
-        loc,
-    };
-}
+import {TokenNode} from "./types";
 
 const print = (
     ast: TokenNode,
@@ -99,7 +29,7 @@ const print = (
             const [numerator, denominator] = ast.children;
             return `(frac@[${loc.path.map(String).join(",")}]:${loc.start}:${
                 loc.end
-            } ${atom.name} ${print(numerator, serialize, indent)} ${print(
+            } ${print(numerator, serialize, indent)} ${print(
                 denominator,
                 serialize,
                 indent,
@@ -114,7 +44,7 @@ const print = (
             const [sub, sup] = ast.children;
             return `(subsup@[${loc.path.map(String).join(",")}]:${loc.start}:${
                 loc.end
-            } ${atom.name} ${sub ? print(sub, serialize, indent) : "_"} ${
+            } ${sub ? print(sub, serialize, indent) : "_"} ${
                 sup ? print(sup, serialize, indent) : "_"
             })`;
         }
@@ -123,7 +53,7 @@ const print = (
             const [lower, upper] = ast.children;
             return `(limits{${inner}}@[${loc.path.map(String).join(",")}]:${
                 loc.start
-            }:${loc.end} ${atom.name} ${print(lower, serialize, indent)} ${
+            }:${loc.end} ${print(lower, serialize, indent)} ${
                 upper ? print(upper, serialize, indent) : "_"
             })`;
         }
@@ -131,7 +61,7 @@ const print = (
             const [index, radicand] = ast.children;
             return `(root@[${loc.path.map(String).join(",")}]:${loc.start}:${
                 loc.end
-            } ${atom.name} ${print(radicand, serialize, indent)} ${
+            } ${print(radicand, serialize, indent)} ${
                 index ? print(index, serialize, indent) : "_"
             })`;
         }
