+++ packages/editor-core/src/token/util.ts
